Move to getProjecs.test.ts
